https-streamlit.io-community-llm-hackathon-2023/requirements.txt
altair==5.1.1
appdirs==1.4.4
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
clarifai==9.8.0
clarifai-grpc==9.8.0
click==8.1.7
cssselect==1.2.0
fake-useragent==1.2.1
gitdb==4.0.10
GitPython==3.1.35
googleapis-common-protos==1.60.0
grpcio==1.58.0
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
numpy==1.25.2
orgparse==0.3.2
packaging==23.1
pandas==2.1.0
parse==1.19.1
Pillow==9.5.0
protobuf==4.24.3
pyarrow==13.0.0
pydeck==0.8.1b0
pyee==8.2.2
Pygments==2.16.1
Pympler==1.0.1
pyppeteer==1.0.2
pyquery==2.0.0
python-dateutil==2.8.2
python-rapidjson==1.10
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
referencing==0.30.2
requests==2.31.0
requests-html==0.10.0
rich==13.4.2
rpds-py==0.10.2
six==1.16.0
smmap==5.0.0
soupsieve==2.5
streamlit==1.23.1
streamlit-diff-viewer==0.0.2
streamlit-js-eval==0.1.5
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.64.1
tritonclient==2.34.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.16
validators==0.22.0
w3lib==2.1.2
watchdog==3.0.0
websockets==10.4
zipp==3.16.2
ratelimit==2.2.1
pyjwt
